Place all pin code in PinUtils.ts
