+++ src/components/MapContainer.vue
@@ -216,388 +216,24 @@ import { Distance } from "@/internal/Distance";
 import DiscoveryDetails from "@/components/DiscoveryDetails.vue";
 import DiscoveryDetailsFullModale from "@/components/DiscoveryDetailsFullModale.vue";
 
-const buildDiscoveryKey = (dType, id) => `${dType}:${id}`;
+import {
+  getPinColors,
+  buildDiscoveryKey,
+  createDefaultPinCanvas,
+  createCircularPhotoPinCanvas,
+  createTargetedPinCanvas,
+  CANVAS_RENDER_SCALE,
+  getCategoryIconName,
+  truncatePinTitle,
+} from "@/internal/PinUtils";
 
-// --- Pin colors per discovery type ---
-// Reads CSS custom properties from :root (defined in GlobalStyle.css)
-// fill = inner/darker, border = outer/paler
-function getCSSVar(name) {
-  return getComputedStyle(document.documentElement).getPropertyValue(name).trim();
-}
-
-function getPinColors(type) {
-  const map = {
-    artwork: { fillStart: getCSSVar('--pin-artwork-fill-start'), fillEnd: getCSSVar('--pin-artwork-fill-end'), border: getCSSVar('--pin-artwork-border') },
-    heritage: { fillStart: getCSSVar('--pin-heritage-fill-start'), fillEnd: getCSSVar('--pin-heritage-fill-end'), border: getCSSVar('--pin-heritage-border') },
-    place: { fillStart: getCSSVar('--pin-place-fill-start'), fillEnd: getCSSVar('--pin-place-fill-end'), border: getCSSVar('--pin-place-border') },
-  };
-  return map[type] ?? map.heritage;
-}
-
-// --- Icon preloading ---
-const iconImages = {}; // name -> HTMLImageElement
-const iconLoadPromises = {};
-
-function preloadSvgIcon(name, path) {
-  if (iconLoadPromises[name]) return iconLoadPromises[name];
-  iconLoadPromises[name] = new Promise((resolve) => {
-    const img = new Image();
-    img.onload = () => { iconImages[name] = img; resolve(img); };
-    img.onerror = () => { iconImages[name] = null; resolve(null); };
-    img.src = path;
-  });
-  return iconLoadPromises[name];
-}
-
-function preloadAllPinIcons() {
-  preloadSvgIcon("targeted", "./assets/drawable/icons/pins/targeted_bookmark.svg");
-  preloadSvgIcon("default", "./assets/drawable/icons/pins/default.svg");
-  preloadSvgIcon("art_public", "./assets/drawable/icons/pins/art_public.svg");
-  preloadSvgIcon("murales", "./assets/drawable/icons/pins/murales.svg");
-  preloadSvgIcon("sculptures", "./assets/drawable/icons/pins/sculptures.svg");
-  preloadSvgIcon("lieux_culturels", "./assets/drawable/icons/pins/lieux_culturels.svg");
-  preloadSvgIcon("bibliotheques", "./assets/drawable/icons/pins/bibliotheques.svg");
-  preloadSvgIcon("patrimoine", "./assets/drawable/icons/pins/patrimoine.svg");
-}
-
-/**
- * Maps a discovery's category/usage string to the corresponding preloaded icon name.
- * For artworks: uses getCategories() (e.g. "Art public", "Murale", "Sculpture")
- * For places: uses getUsages() (e.g. "Bibliothèque", "Maison de la culture")
- * For heritage: always "patrimoine"
- */
-function getCategoryIconName(discovery) {
-  if (!discovery) return "default";
-
-  const dType = discovery.dType;
-
-  if (dType === "heritage") return "patrimoine";
-
-  // Get the first category/usage string
-  let rawCategory = "";
-  if (dType === "artwork" && typeof discovery.getCategories === "function") {
-    rawCategory = discovery.getCategories("fr");
-  } else if (dType === "place" && typeof discovery.getUsages === "function") {
-    rawCategory = discovery.getUsages("fr");
-  }
-
-  // Take only the first value if comma-separated
-  const first = rawCategory.split(",")[0].trim().toLowerCase();
-  if (!first) {
-    // No category found — use type-specific default icon
-    if (dType === "artwork") return "art_public";
-    if (dType === "place") return "lieux_culturels";
-    return "default";
-  }
-
-  // Map to icon name
-  if (first.includes("art public")) return "art_public";
-  if (first.includes("murale")) return "murales";
-  if (first.includes("sculpture")) return "sculptures";
-  if (first.includes("biblioth")) return "bibliotheques";
-  if (first.includes("lieu") || first.includes("maison de la culture") || first.includes("centre") || first.includes("galerie") || first.includes("mus")) return "lieux_culturels";
-  if (first.includes("patrimoine")) return "patrimoine";
-
-  // Unrecognized category — use type-specific default
-  if (dType === "artwork") return "art_public";
-  if (dType === "place") return "lieux_culturels";
-  return "default";
-}
-
-// Start preloading immediately
-preloadAllPinIcons();
-
-// --- Collected photo pin caches ---
-const collectedPhotoImgCache = {};
-const collectedPhotoPinCache = {};
-
-// --- Targeted pin cache ---
+// Use cached objects/methods to save on rendering time
+const collectedPhotoImgCache = {}; // "type:id" -> HTMLImageElement
+const collectedPhotoPinCache = {}; // "type:id:size" -> canvas
 const targetedPinCache = {}; // "type:title:size" -> canvas
-
-// --- Default pin cache ---
 const defaultPinCache = {}; // "type:size" -> canvas
 
-// Fixed render scale for crisp canvases on all screens (DPR-independent)
-const CANVAS_RENDER_SCALE = 2;
-
-function createCircularPhotoPinCanvas(img, type, size = 56, isSelected = false) {
-  const colors = getPinColors(type);
-  const ringWidth = 2;
-  const totalSize = size + ringWidth * 2;
-  const pointerHeight = 10;
-  // +4 to make room for ground shadow
-  const shadowPaddingBottom = isSelected ? 4 : 0;
-
-  const canvas = document.createElement("canvas");
-  canvas.width = totalSize * CANVAS_RENDER_SCALE;
-  canvas.height = (totalSize + pointerHeight + shadowPaddingBottom) * CANVAS_RENDER_SCALE;
-  // Store logical size for OpenLayers imgSize
-  canvas._logicalWidth = totalSize;
-  canvas._logicalHeight = totalSize + pointerHeight + shadowPaddingBottom;
-
-  const ctx = canvas.getContext("2d");
-  ctx.scale(CANVAS_RENDER_SCALE, CANVAS_RENDER_SCALE);
-
-  const cx = totalSize / 2;
-  const cy = totalSize / 2;
-  const photoRadius = size / 2;
-  const tipY = cy + photoRadius + ringWidth + pointerHeight - 2;
-
-  if (isSelected) {
-    // Ground shadow ellipse
-    ctx.fillStyle = "black";
-    ctx.beginPath();
-    ctx.ellipse(cx, tipY, 14, 3, 0, 0, Math.PI * 2);
-    ctx.fill();
-  }
-
-  // Shadow
-  ctx.shadowColor = "rgba(0, 0, 0, 0.3)";
-  ctx.shadowBlur = 6;
-  ctx.shadowOffsetY = 2;
-
-  // Gradient ring (single ring, no outer border)
-  const ringGrad = ctx.createLinearGradient(cx, cy - (photoRadius + ringWidth), cx, cy + (photoRadius + ringWidth));
-  ringGrad.addColorStop(0, colors.fillStart);
-  ringGrad.addColorStop(1, colors.fillEnd);
-  ctx.beginPath();
-  ctx.arc(cx, cy, photoRadius + ringWidth, 0, Math.PI * 2);
-  ctx.fillStyle = ringGrad;
-  ctx.fill();
-
-  // Gradient pointer
-  ctx.beginPath();
-  ctx.moveTo(cx - 6, cy + photoRadius + ringWidth - 2);
-  ctx.lineTo(cx, tipY);
-  ctx.lineTo(cx + 6, cy + photoRadius + ringWidth - 2);
-  ctx.closePath();
-  ctx.fillStyle = colors.fillEnd;
-  ctx.fill();
-
-  // Reset shadow
-  ctx.shadowColor = "transparent";
-  ctx.shadowBlur = 0;
-  ctx.shadowOffsetY = 0;
-
-  // Clip circle and draw photo
-  ctx.save();
-  ctx.beginPath();
-  ctx.arc(cx, cy, photoRadius, 0, Math.PI * 2);
-  ctx.clip();
-
-  const imgSize = Math.min(img.width, img.height);
-  const sx = (img.width - imgSize) / 2;
-  const sy = (img.height - imgSize) / 2;
-  ctx.drawImage(img, sx, sy, imgSize, imgSize, cx - photoRadius, cy - photoRadius, size, size);
-  ctx.restore();
-
-  return canvas;
-}
-
-/**
- * Creates a targeted/bookmarked pin: rounded pill with bookmark icon + title + pointer at bottom.
- * Matches Figma design.
- */
-function createTargetedPinCanvas(title, colors, isSelected = false) {
-  const paddingX = 10;
-  const paddingY = 6;
-  const iconSize = 12;
-  const iconGap = 5;
-  const fontSize = 12;
-  const pointerHeight = 8;
-  const borderRadius = 14;
-
-  // Measure text
-  const measureCanvas = document.createElement("canvas");
-  const measureCtx = measureCanvas.getContext("2d");
-  measureCtx.font = `700 ${fontSize}px Arial`;
-  const displayTitle = truncatePinTitle(title);
-  const textWidth = measureCtx.measureText(displayTitle).width;
-
-  const pillWidth = paddingX + iconSize + iconGap + textWidth + paddingX;
-  const pillHeight = paddingY * 2 + fontSize + 2;
-  const totalWidth = pillWidth;
-  // +4 to make room for ground shadow
-  const shadowPaddingBottom = isSelected ? 4 : 0;
-  const totalHeight = pillHeight + pointerHeight + shadowPaddingBottom;
-
-  const canvas = document.createElement("canvas");
-  canvas.width = totalWidth * CANVAS_RENDER_SCALE;
-  canvas.height = totalHeight * CANVAS_RENDER_SCALE;
-  canvas._logicalWidth = totalWidth;
-  canvas._logicalHeight = totalHeight;
-
-  const ctx = canvas.getContext("2d");
-  ctx.scale(CANVAS_RENDER_SCALE, CANVAS_RENDER_SCALE);
-
-  const cx = totalWidth / 2;
-  const tipY = pillHeight + pointerHeight - 1;
-
-  if (isSelected) {
-    // Ground shadow ellipse
-    ctx.fillStyle = "black";
-    ctx.beginPath();
-    ctx.ellipse(cx, tipY, 14, 3, 0, 0, Math.PI * 2);
-    ctx.fill();
-  }
-
-  // Shadow
-  ctx.shadowColor = "rgba(0, 0, 0, 0.25)";
-  ctx.shadowBlur = 4;
-  ctx.shadowOffsetY = 1;
-
-  // Pill gradient (vertical)
-  const pillGrad = ctx.createLinearGradient(0, 0, 0, pillHeight);
-  pillGrad.addColorStop(0, colors.fillStart);
-  pillGrad.addColorStop(1, colors.fillEnd);
-
-  // Rounded pill
-  ctx.beginPath();
-  ctx.moveTo(borderRadius, 0);
-  ctx.lineTo(pillWidth - borderRadius, 0);
-  ctx.quadraticCurveTo(pillWidth, 0, pillWidth, borderRadius);
-  ctx.lineTo(pillWidth, pillHeight - borderRadius);
-  ctx.quadraticCurveTo(pillWidth, pillHeight, pillWidth - borderRadius, pillHeight);
-  ctx.lineTo(borderRadius, pillHeight);
-  ctx.quadraticCurveTo(0, pillHeight, 0, pillHeight - borderRadius);
-  ctx.lineTo(0, borderRadius);
-  ctx.quadraticCurveTo(0, 0, borderRadius, 0);
-  ctx.closePath();
-  ctx.fillStyle = pillGrad;
-  ctx.fill();
-
-  // Pointer (use gradient end color)
-  ctx.beginPath();
-  ctx.moveTo(cx - 6, pillHeight - 1);
-  ctx.lineTo(cx, tipY);
-  ctx.lineTo(cx + 6, pillHeight - 1);
-  ctx.closePath();
-  ctx.fillStyle = colors.fillEnd;
-  ctx.fill();
-
-  // Reset shadow
-  ctx.shadowColor = "transparent";
-  ctx.shadowBlur = 0;
-  ctx.shadowOffsetY = 0;
-
-  // Draw bookmark icon
-  const bookmarkIcon = iconImages["targeted"];
-  if (bookmarkIcon) {
-    const iconY = (pillHeight - iconSize) / 2;
-    ctx.drawImage(bookmarkIcon, paddingX, iconY, iconSize, iconSize * (18 / 13));
-  }
-
-  // Draw title text
-  ctx.font = `700 ${fontSize}px Arial`;
-  ctx.fillStyle = "#1F1F1F";
-  ctx.textBaseline = "middle";
-  ctx.fillText(displayTitle, paddingX + iconSize + iconGap, pillHeight / 2 + 1);
-
-  return canvas;
-}
-
-/**
- * Creates a default pin: teardrop/balloon shape with an icon inside.
- * Matches Figma design.
- */
-function createDefaultPinCanvas(type, categoryIcon = "default", isSelected = false) {
-  const colors = getPinColors(type);
-  const borderWidth = 1;
-  const innerRadius = 12;
-  const outerRadius = innerRadius + borderWidth;
-  const pointerHeight = 10;
-  const totalSize = (outerRadius + 2) * 2; // +2 for shadow margin
-  // +4 to make room for ground shadow
-  const shadowPaddingBottom = isSelected ? 4 : 0;
-  const totalHeight = totalSize + pointerHeight + shadowPaddingBottom;
-
-  const canvas = document.createElement("canvas");
-  canvas.width = totalSize * CANVAS_RENDER_SCALE;
-  canvas.height = totalHeight * CANVAS_RENDER_SCALE;
-  canvas._logicalWidth = totalSize;
-  canvas._logicalHeight = totalHeight;
-
-  const ctx = canvas.getContext("2d");
-  ctx.scale(CANVAS_RENDER_SCALE, CANVAS_RENDER_SCALE);
-
-  const cx = totalSize / 2;
-  const cy = outerRadius + 1;
-  const tipY = cy + outerRadius + pointerHeight - 2;
-
-  if (isSelected) {
-    // Ground shadow ellipse
-    ctx.fillStyle = "black";
-    ctx.beginPath();
-    ctx.ellipse(cx, tipY, 14, 3, 0, 0, Math.PI * 2);
-    ctx.fill();
-  }
-
-  // Shadow
-  ctx.shadowColor = "rgba(0, 0, 0, 0.25)";
-  ctx.shadowBlur = 3;
-  ctx.shadowOffsetY = 1;
-
-  // Outer pale border circle
-  ctx.beginPath();
-  ctx.arc(cx, cy, outerRadius, 0, Math.PI * 2);
-  ctx.fillStyle = colors.border;
-  ctx.fill();
-
-  // Outer pale pointer
-  ctx.beginPath();
-  ctx.moveTo(cx - 6, cy + outerRadius - 2);
-  ctx.lineTo(cx, tipY);
-  ctx.lineTo(cx + 6, cy + outerRadius - 2);
-  ctx.closePath();
-  ctx.fillStyle = colors.border;
-  ctx.fill();
-
-  // Reset shadow for inner fill
-  ctx.shadowColor = "transparent";
-  ctx.shadowBlur = 0;
-  ctx.shadowOffsetY = 0;
-
-  // Inner darker fill circle (vertical gradient)
-  const innerGrad = ctx.createLinearGradient(cx, cy - innerRadius, cx, cy + innerRadius);
-  innerGrad.addColorStop(0, colors.fillStart);
-  innerGrad.addColorStop(1, colors.fillEnd);
-  ctx.beginPath();
-  ctx.arc(cx, cy, innerRadius, 0, Math.PI * 2);
-  ctx.fillStyle = innerGrad;
-  ctx.fill();
-
-  // Inner darker pointer (use gradient end color)
-  ctx.beginPath();
-  ctx.moveTo(cx - 4, cy + innerRadius - 1);
-  ctx.lineTo(cx, cy + innerRadius + pointerHeight - 4);
-  ctx.lineTo(cx + 4, cy + innerRadius - 1);
-  ctx.closePath();
-  ctx.fillStyle = colors.fillEnd;
-  ctx.fill();
-
-  // Draw category icon inside the circle
-  const icon = iconImages[categoryIcon] || iconImages["default"];
-  if (icon) {
-    const maxIconSize = 13;
-    const imgWidth = icon.width || 1;
-    const imgHeight = icon.height || 1;
-    const imgAspect = imgWidth / imgHeight;
-    
-    let drawWidth = maxIconSize;
-    let drawHeight = maxIconSize;
-
-    if (imgAspect > 1) {
-      drawHeight = maxIconSize / imgAspect;
-    } else {
-      drawWidth = maxIconSize * imgAspect;
-    }
-
-    ctx.drawImage(icon, cx - drawWidth / 2, cy - drawHeight / 2, drawWidth, drawHeight);
-  }
-
-  return canvas;
-}
-
+// --- Pin colors per discovery type ---
 function getDiscoveryLocation(discovery) {
   if (!discovery) return null;
 
@@ -661,12 +297,6 @@ function insertAllPins(
     });
     destinationLayer.getSource().addFeature(feature);
   }
-}
-
-function truncatePinTitle(title) {
-  const maxChars = 16;
-  if (!title || typeof title !== "string") return "";
-  return title.length > maxChars ? `${title.slice(0, maxChars)}...` : title;
 }
 
 function insertAllPolygons(destinationLayer, discoveryList) {
